Remove jQuery dependency; closes #3.
jQuery was only being used to search for the relevant block
to pull out of the transcluded clone. Change to simply iterate
over clone "by hand", searching for the node with the correct
name. This means that all named blocks must be immediate children
of the directive application; this doesn't seem like too much
of a restriction, as any other nodes are ignored anyway.